Fix recipe file sorting bug
Problem: the files containing the recipe data were being processed in the wrong order. The JSON  output didn't match the ebook.

Cause: the File API upload feature doesn't maintain the file order of the directory.

Solution: add a sorting function to arrange the files in the same order as the ebook.